+++ services/bss/src/main/java/com/huaweicloud/sdk/bss/v2/model/OrderRefundInfoV2.java
@@ -3,6 +3,7 @@ package com.huaweicloud.sdk.bss.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 /**
@@ -18,7 +19,7 @@ public class OrderRefundInfoV2 {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "amount")
 
-    private Double amount;
+    private BigDecimal amount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "measure_id")
@@ -77,7 +78,7 @@ public class OrderRefundInfoV2 {
         this.id = id;
     }
 
-    public OrderRefundInfoV2 withAmount(Double amount) {
+    public OrderRefundInfoV2 withAmount(BigDecimal amount) {
         this.amount = amount;
         return this;
     }
@@ -86,11 +87,11 @@ public class OrderRefundInfoV2 {
      * 金额。 金额为负数，表示退订金额。金额为正数，表示已消费金额或收取的退订手续费。
      * @return amount
      */
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
